Fix missing LaunchedEffect import and unify Long arithmetic in score calculations
- PerformanceGraphsActivity: LaunchedEffect was unresolved after the runtime import
  was narrowed to individual imports
- Player/MatchLog/DailyInput: use explicit Long literals in score math for consistency

+++ app/src/main/java/com/zerostress/manager/models/MatchLog.kt
@@ -16,7 +16,7 @@ data class MatchLog(
     var date: Long = 0L
 ) {
     val score: Long
-        get() = (kills * 10 + damage / 100 + (if (win) 200 else 0)).toLong()
+        get() = (kills * 10L + damage / 100 + (if (win) 200L else 0L))
 
     companion object {
         fun create(
